PEAUTY-18 : pawfit -> peauty
