+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
     ports:
       - "3306:3306"
     environment:
-      MYSQL_DATABASE: pawfit
-      MYSQL_USER: pawfit
+      MYSQL_DATABASE: Peauty
+      MYSQL_USER: Peauty
       MYSQL_PASSWORD: 1234
       MYSQL_ROOT_PASSWORD: 1234
