--- pawfit-aws/src/main/java/com/pawfit/aws/S3Config.java
@@ -1,4 +1,0 @@
-package com.pawfit.aws;
-
-public class S3Config {
-}
