--- pawfit-customer/build.gradle
@@ -1,19 +1,0 @@
-bootJar {
-    enabled = true
-}
-
-jar {
-    enabled = false
-}
-
-dependencies {
-    implementation(project(':pawfit-persistence'))
-    implementation(project(':pawfit-domain'))
-    implementation(project(':pawfit-logging'))
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    //implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.security:spring-security-test'
-}
