--- pawfit-customer/src/main/java/com/pawfit/customer/PawfitCustomerApplication.java
@@ -1,13 +1,0 @@
-package com.pawfit.customer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PawfitCustomerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PawfitCustomerApplication.class, args);
-    }
-
-}
