--- pawfit-customer/src/test/java/com/pawfit/customer/PawfitCustomerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.pawfit.customer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PawfitCustomerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
