--- pawfit-designer/src/main/java/com/pawfit/designer/PawfitDesignerApplication.java
@@ -1,13 +1,0 @@
-package com.pawfit.designer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PawfitDesignerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PawfitDesignerApplication.class, args);
-	}
-
-}
