--- pawfit-designer/src/main/resources/application.yml
@@ -1,18 +1,0 @@
-server:
-  port: 8080
-  shutdown: graceful
-
-spring:
-  application:
-    name: pawfit-designer
-  profiles:
-    active: local
-    include:
-      - persistence
-      - aws
-  jackson:
-    default-property-inclusion: non_null
-  config:
-    import:
-      - classpath:properties/oauth.yml
-      - classpath:properties/jwt.yml # TODO 모듈로 빼기
