--- pawfit-designer/src/test/java/com/pawfit/designer/PawfitDesignerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.pawfit.designer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PawfitDesignerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
