--- pawfit-domain/src/main/java/com/pawfit/domain/customer/Customer.java
@@ -1,4 +1,0 @@
-package com.pawfit.domain.customer;
-
-public record Customer(Long id, String name) {
-}
