--- pawfit-persistence/src/main/java/com/pawfit/persistence/JpaConfig.java
@@ -1,4 +1,0 @@
-package com.pawfit.persistence;
-
-public class JpaConfig {
-}
