+++ peauty-aws/src/main/java/com/peauty/aws/S3Config.java
@@ -1,0 +1,4 @@
+package com.peauty.aws;
+
+public class S3Config {
+}
