+++ peauty-aws/src/main/resources/application-aws.yml
@@ -1,0 +1,5 @@
+Peauty:
+  aws:
+    region: "test-region"
+    s3:
+      bucket-name: "test-bucket"
