+++ peauty-customer/src/main/java/com/peauty/customer/PeautyCustomerApplication.java
@@ -1,0 +1,13 @@
+package com.peauty.customer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PeautyCustomerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PeautyCustomerApplication.class, args);
+    }
+
+}
