+++ peauty-customer/src/test/java/com/peauty/customer/PeautyCustomerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.peauty.customer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PeautyCustomerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
