+++ peauty-designer/build.gradle
@@ -1,0 +1,19 @@
+bootJar {
+	enabled = true
+}
+
+jar {
+	enabled = false
+}
+
+dependencies {
+	implementation(project(':peauty-persistence'))
+	implementation(project(':peauty-domain'))
+	implementation(project(':peauty-logging'))
+	implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+	//implementation 'org.springframework.boot:spring-boot-starter-security'
+	implementation 'org.springframework.boot:spring-boot-starter-validation'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testImplementation 'org.springframework.security:spring-security-test'
+}
