+++ peauty-designer/src/main/java/com/peauty/designer/PeautyDesignerApplication.java
@@ -1,0 +1,13 @@
+package com.peauty.designer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PeautyDesignerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PeautyDesignerApplication.class, args);
+	}
+
+}
