+++ peauty-designer/src/main/resources/application.yml
@@ -1,0 +1,18 @@
+server:
+  port: 8080
+  shutdown: graceful
+
+spring:
+  application:
+    name: Peauty-designer
+  profiles:
+    active: local
+    include:
+      - persistence
+      - aws
+  jackson:
+    default-property-inclusion: non_null
+  config:
+    import:
+      - classpath:properties/oauth.yml
+      - classpath:properties/jwt.yml # TODO 모듈로 빼기
