+++ peauty-designer/src/main/resources/properties/oauth.yml
