+++ peauty-designer/src/test/java/com/peauty/designer/PeautyDesignerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.peauty.designer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PeautyDesignerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
