+++ peauty-domain/src/main/java/com/peauty/domain/customer/Customer.java
@@ -1,0 +1,4 @@
+package com.peauty.domain.customer;
+
+public record Customer(Long id, String name) {
+}
