+++ peauty-persistence/src/main/java/com/peauty/persistence/CustomerEntity.java
@@ -1,0 +1,4 @@
+package com.peauty.persistence;
+
+public class CustomerEntity {
+}
