+++ peauty-persistence/src/main/java/com/peauty/persistence/JpaConfig.java
@@ -1,0 +1,4 @@
+package com.peauty.persistence;
+
+public class JpaConfig {
+}
