+++ settings.gradle
@@ -1,8 +1,8 @@
-rootProject.name = 'pawfit'
+rootProject.name = 'peauty'
 
-include ':pawfit-customer'
-include ':pawfit-designer'
-include ':pawfit-domain'
-include ':pawfit-aws'
-include ':pawfit-persistence'
-include ':pawfit-logging'
+include ':peauty-customer'
+include ':peauty-designer'
+include ':peauty-domain'
+include ':peauty-aws'
+include ':peauty-persistence'
+include ':peauty-logging'
